+++ modules/core/server-ts/graphql/index.ts
@@ -1,3 +1,3 @@
-export { default as createApolloServer } from './graphql';
+export { default as createApolloServer } from './apolloServer';
 export { createSchema } from './schema';
 export { default as addGraphQLSubscriptions, onAppDispose } from './subscriptions';
